Change a place where we convert dict to our custom dict wrapper for user friendly-errors.
This way we ensure self.config is always indeed an instance of
PackConfigDict class.
